+++ resources/src/filters/craft.js
@@ -1,0 +1,4 @@
+export function t (message, params = null) {
+    // @ts-ignore
+    return window.Craft.t('simplemap', message, params);
+}